[add] volumetic mesh structure & other gui stuff
in short, no progress for now
