--- Assets/New Material.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d346aef49016de44eba2f382661d7438
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
