+++ Assets/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d42dace3c2664b429125113ff8211e2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
