+++ Assets/Script/GlassBreakController.cs
@@ -1,0 +1,25 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class GlassBreakController : MonoBehaviour
+{
+    public Slider energySlider;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void startBreaking() {
+        
+    }
+}
