+++ Assets/Script/Splitter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ffa5fbb8f803a5748a54f2336180c732
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
