--- Assets/ShaderBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3ec32c305bbf6fd4296fe5e1cf53eca5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
