+++ Assets/Skybox.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef8b4bc8f45e94b848276317234204b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
