+++ Assets/Skybox/Sunset Skybox.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa05c057326984739bae44f9be0ed44b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
