+++ Assets/Structure.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d22ff350f811547d3a6c3b6b1bb4c4a1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
