+++ Assets/Structure/VolumeticMesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2f5c8f7cb507460796d332ee273523b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
