--- Assets/ToolFunction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a34e94cb0edc5249ac59d0e38606c4a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
